reorganised the statdb & authentication package and update some notes
