+++ core/vm/interface.go
@@ -62,8 +62,6 @@ type StateDB interface {
 	AddPreimage(common.Hash, []byte)
 
 	ForEachStorage(common.Address, func(common.Hash, common.Hash) bool)
-
-	CheckMultiAccountSig(*types.Transaction, int, common.Address) error
 }
 
 // CallContext provides a basic interface for the EVM calling conventions. The EVM EVM
